[decompiler] Use meters, degrees, and seconds (#689)
* use meters degrees and seconds

* update changelog

+++ decompiler/IR2/AtomicOpTypeAnalysis.cpp
@@ -557,7 +557,7 @@ TP_Type SimpleExpression::get_type_int2(const TypeState& input,
   if (m_kind == Kind::ADD && tc(dts, TypeSpec("structure"), arg0_type) &&
       arg1_type.is_integer_constant()) {
     auto type_info = dts.ts.lookup_type(arg0_type.typespec());
-    if (type_info->get_size_in_memory() == arg1_type.get_integer_constant()) {
+    if ((u64)type_info->get_size_in_memory() == arg1_type.get_integer_constant()) {
       return TP_Type::make_from_ts(arg0_type.typespec());
     }
   }
